arch-co-pilot-backend/applications/arch_copilot/requirements.txt
uvicorn
fastapi
boto3
pandas
uuid
pymupdf
Pillow
pypdfium2
PyYAML
opencv-python
moviepy
ffmpeg-python
imageio